Update LICENSE
I actually don't believe I was ever actually supposed to be able to make this MIT.
Technically, this library uses a network service which uses the AGPL license, and when I read through,
it seemed like it required that anything that uses the networks services provided by the terms of that license to use that license, so anyway. This is that change.